distributed_loader: Don't check for subdir absence twice
There are two calls to file_exists() in dist.loader. One is done when
starting a subdir for population, the other one is when the subdir is
used for population. The latter call is excessive and can be replaced
with simpler check for the sstable_dir instance presense in the
population metadata.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>