sstable_directory: Construct it with table location
Currently the directory is constructed with full sstable path, but table
is going to stop knowing it. To prepare for that, the sstable directory
should be constructed from the table location.

To get the full path it has to talk to sstables-manager to get full path
the way it does. That's somewhat temporary, sstable directory needs full
path to do three things:

- make sstables
- list sstables
- process pending deletion

While the first item can already be patched to use location, the last
two still need more care.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>